pyproject.toml
[project]
name = "paid-python"

[tool.poetry]
name = "paid-python"
version = "0.1.0"
description = ""
readme = "README.md"
authors = []
keywords = []

classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
packages = [
    { include = "paid", from = "src"}
]

[project.urls]
Repository = 'https://github.com/paid-ai/paid-python'

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
httpx = ">=0.21.0"
pydantic = ">=1.9.0"
pydantic-core = ">=2.18.0"
typing_extensions = ">= 4.0.0"
opentelemetry-api = ">=1.23.0"
opentelemetry-sdk = ">=1.23.0"
opentelemetry-exporter-otlp-proto-http = ">=1.23.0"
mutagen = ">=1.47.0"
python-dotenv = ">=0.15.0"
opentelemetry-instrumentation-anthropic = ">=0.47.0"
opentelemetry-instrumentation-google-generativeai = ">=0.47.0"
opentelemetry-instrumentation-openai = ">=0.47.0"
openinference-instrumentation-openai-agents = ">=1.0.0"
openinference-instrumentation-bedrock = ">=0.1.0"

[tool.poetry.group.dev.dependencies]
mypy = "==1.13.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.23.5"
python-dateutil = "^2.9.0"
types-python-dateutil = "^2.9.0.20240316"
ruff = "==0.11.5"

# Peer-dependencies, but useful to have for development
mistralai = ">=1.9.0"
anthropic = ">=0.57.0"
llama-index-core = ">=0.13.0"
llama-index-llms-openai = ">=0.5.0"
google-genai = ">=1.19.0"
openai-agents = ">=0.2.0"
langchain-core = ">=0.1.0"
openai = ">=1.12.0"

[tool.pytest.ini_options]
testpaths = [ "tests" ]
asyncio_mode = "auto"

[tool.mypy]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "anthropic",
    "anthropic.*",
    "openai",
    "mistralai",
    "mistralai.*",
    "llama_index.*",
    "google.*",
    "agents",
    "agents.*",
    "langchain_core.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
]
ignore = [
    "E402",  # Module level import not at top of file
    "E501",  # Line too long
    "E711",  # Comparison to `None` should be `cond is not None`
    "E712",  # Avoid equality comparisons to `True`; use `if ...:` checks
    "E721",  # Use `is` and `is not` for type comparisons, or `isinstance()` for insinstance checks
    "E722",  # Do not use bare `except`
    "E731",  # Do not assign a `lambda` expression, use a `def`
    "F821",  # Undefined name
    "F841"   # Local variable ... is assigned to but never used
]

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
